requirements (1).txt
fastapi
uvicorn
pydantic
openai
requests
python-dotenv
streamlit